requirements.txt
pyserial
docker
pyaml-env==1.1.3
pynacl==1.4.0
sphinx==4.3.2
